+++ src/controllers/ClearingTransaction.ts
@@ -4,8 +4,9 @@ import { Request, Response, NextFunction } from 'express';
 import { createClearingTransactionAndBroadcast } from '../functions/builders/clearing-tx-builder';
 import { loadConfig, getPrivateKeyById } from '../functions/config';
 import logger from '../functions/logger';
-import { ClearingTransactionPayload, ClearingTransactionResponse } from '../types/api/clearingTransaction';
 import { ClearingTransactionPayloadSchema } from '../validator/schemas';
+
+import type { ClearingTransactionPayload, ClearingTransactionResponse } from '../types/api/clearingTransaction';
 
 const config = loadConfig();
 const keyring = new Keyring({ type: 'sr25519' });
