+++ src/types/api/common.ts
@@ -1,51 +1,3 @@
-import type { ActionType } from './actions';
-
-/**
- * The origin type for each Action in CT
- */
-export type CTActionOrigin =
-  | { AppAgentId: any }
-  | { AppAgentAddress: string }
-  | { TransactionalAddressId: any }
-  | { TransactionalAddress: any }
-  | { NamedAddress: any }
-  | { NamedAddressName: any };
-
-/**
- * Represents an action in a clearing transaction.
- */
-export type CTAction = [
-  // The origin of the action
-  CTActionOrigin,
-  /**
-   * The call to be made by the action
-   * To take advantage of txwrapper methods, this could be UnsignedTransaction.method.
-   */
-  { callIndex?: string; args?: string } | string,
-];
-
-/**
- * A nested array of actions (Vec<Vec<Action>> in rust).
- */
-export type CTAtomicActions = CTAction[][];
-
-export type CTAtomicActionGeneric<ActionTypeName extends ActionType, ActionArgs> = {
-  /**
-   * The type of action to be performed
-   */
-
-  actionType: ActionTypeName;
-  /**
-   * The origin responsible for the transaction.
-   */
-
-  origin: CTActionOrigin;
-  /**
-   * The arguments for the action.
-   */
-  arguments: ActionArgs;
-};
-
 // Common blockchain types
 
 /**
